+++ Runtime/Scripts/DefaultDeferAgent.cs
@@ -20,12 +20,21 @@ namespace GLTFast {
     
     /// <summary>
     /// To be added to a GameObject along a default <see cref="IDeferAgent"/>.
-    /// Will un-register itself from GltfImport when it's destroyed.
+    /// Will (un)register it as GltfImport's default when it's enabled or disabled.
     /// </summary>
     [RequireComponent(typeof(IDeferAgent))]
+    [DefaultExecutionOrder(-1)]
     class DefaultDeferAgent : MonoBehaviour {
-        void OnDestroy() {
-            IDeferAgent deferAgent = GetComponent<IDeferAgent>();
+
+        void OnEnable() {
+            var deferAgent = GetComponent<IDeferAgent>();
+            if (deferAgent != null) {
+                GltfImport.SetDefaultDeferAgent(deferAgent);
+            }
+        }
+
+        void OnDisable() {
+            var deferAgent = GetComponent<IDeferAgent>();
             if (deferAgent != null) {
                 GltfImport.UnsetDefaultDeferAgent(deferAgent);
             }